+++ server/internal/http/admin/dto/log.go
@@ -8,7 +8,7 @@ type LogTypeListReq struct {
 type LogListReq struct {
 	Page      int    `form:"page" validate:"min:1"`
 	PageSize  int    `form:"page_size" validate:"min:1|max:500"`
-	TypeId    int64  `form:"type_id"`
+	TypeId    int32  `form:"type_id"`
 	StartTime string `form:"start_time"`
 	EndTime   string `form:"end_time"`
 }
