+++ server/internal/orm/model/menu_api.gen.go
@@ -10,9 +10,9 @@ const TableNameMenuAPI = "menu_api"
 
 // MenuAPI 菜单权限接口
 type MenuAPI struct {
-	ID     int64  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
-	AppID  int64  `gorm:"column:app_id;not null;comment:应用ID" json:"app_id"`   // 应用ID
-	MenuID int64  `gorm:"column:menu_id;not null;comment:菜单ID" json:"menu_id"` // 菜单ID
+	ID     int32  `gorm:"column:id;primaryKey;autoIncrement:true" json:"id"`
+	AppID  int32  `gorm:"column:app_id;not null;comment:应用ID" json:"app_id"`   // 应用ID
+	MenuID int32  `gorm:"column:menu_id;not null;comment:菜单ID" json:"menu_id"` // 菜单ID
 	Path   string `gorm:"column:path;not null;comment:API路由地址" json:"path"`    // API路由地址
 	Tag    string `gorm:"column:tag;not null;comment:标识" json:"tag"`           // 标识
 }
